Add rudamentary server-side bot support
Adds a "/shadow" command which replaces your current player with a
server-side version. Currently these are a bit bugged, as they can
fly in the air if blocks are removed and do not take proper damage.

However, I believe if player intention is not to abuse this it can
be very useful to save on energy, instead of keeping computers on.